requirements-prod.txt
# Production requirements for CreativeIQ
# Core dependencies (frozen versions for stability)

# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# AI/ML Core
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
huggingface-hub==0.19.4
accelerate==0.24.1

# Computer Vision
opencv-python-headless==4.8.1.78  # Headless for production
Pillow==10.1.0
scikit-image==0.22.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3
joblib==1.3.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# Image Processing & Analysis
pytesseract==0.3.10
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.0
httpx==0.25.2

# Production extras
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Performance
orjson==3.9.10
ujson==5.8.0